requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
pydantic
ingredient-slicer
